+++ packages/svelte/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* `<Link>` uses `history.navigate`, not `history.update`.
+
 ## 1.0.0-alpha.3
 
 * Only call `history.update` for left, non-modified `<Link>` clicks.